+++ project/dayoung-onboarding/src/main/kotlin/dayoung/onboarding/domain/user/User.kt
@@ -1,4 +1,4 @@
-package dayoung.onboarding.domain
+package dayoung.onboarding.domain.user
 
 import jakarta.persistence.*
